feat: mock 모드 해외거래 차단 + USD 표기 수정 + KIS Builder 전략 연동
- mock 모드에서 해외주식 매수·매도·스크리닝 전면 차단
- 해외 종목 매도 시 sell_overseas() 라우팅 수정 (sell → sell_overseas)
- OrderRecord.price/amount int→float (USD 소수점 보존)
- 주문내역·보유종목 해외 종목 USD 표기 ($x.xx / 원 구분)
- update_filled_price: 해외 종목 yfinance fallback 추가
- KisBuilderStrategy 추가 — Strategy Builder(8001) HTTP 연동
- config에 strategy_id / kis_builder_strategy 기본값 추가
- 웹 config 페이지 투자한도 max를 net_asset 기준으로 동적 설정
- KIS 500 에러 전체 재시도 (기존: EGW00201만)
- 보유종목 market 필드 추가 (UniverseSymbol DB 조인)
- scripts/start_strategy_builder.sh 추가

+++ web/api/models.py
@@ -32,8 +32,8 @@ class OrderRecord(SQLModel, table=True):
     name: str = ""
     side: str  # buy / sell
     qty: int
-    price: int
-    amount: int  # qty * price
+    price: float
+    amount: float  # qty * price
     reason: str = ""
     strategy: str = ""
     created_at: datetime = Field(default_factory=datetime.now)
